Factor toUserFriendlyAddress out into AddressUtils
